Added an optimization, called removeUnnecessaryAutoCopyCalls, that attempts to remove matching autoCopy and autoDestroy calls (matching in the sense that they cancel one another out) within a basic block. This happens quite often because autoCopy calls are moved into functions when a returned record is changed into a pass-by-reference record.
This optimization is restricted to the autoCopy and autoDestroy
functions for arrays and domains to avoid the question of side effects
in the normal record case.  Fortunately, this optimization is most
important for arrays and domains due to the high overhead of the
reference counting.  This restriction is implemented via new flags
"removable auto copy" and "removable auto destroy" which mark the
autoCopy and autoDestroy functions as candidates for the above
optimization (respectively).

The optimization is local to basic blocks and proceeds as follows:

  for statements in basic block do
    if statement is an autoCopy call then
      add actual to map from actual to autoCopy call
    if statement is an autoDestroy call then
      if actual is in map from actual to autoCopy call then
        remove autoCopy and autoDestroy calls
    for all symbols used/defined in statement do
      remove symbol from map from symbol to autoCopy call

Additionally, to make the optimization more effective, an equivalence
map is maintained between variables that are made equivalent via a
primitive move.  The equivalence map is used to increase the matches
between autoCopy and autoDestroy calls by making the optimization
apply if the actual to the autoCopy is equivalent to the actual to the
autoDestroy.  Entries in this equivalence map are invalidated if they
are found to be used or defined in other statements.

This optimization improves the performance on some tests.  For
example, the execution time of test/npb/cg/bradc/cg-sparse.chpl when
compiled with --no-local and --no-checks improves further (see
revision 15921 for more information) with this commit:

  Before this commit    0m27.274s on chevy-lnx
  After this commit     0m20.581s on chevy-lnx

As mentioned in revision 15921, before revision 15875 (initial support
for inserting calls to the destructors of records via autoCopy and
autoDestroy), the time was even better: 0m10.473s on chevy-lnx.  So
there remains more optimization work to do.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15923 3a8e244f-b0f2-452b-bcba-4c88e055c3ca