Fix a couple of problems with the new generic tasking implementation.
First, valgrind testing revealed that the thread private serial_state
value for the main thread never was getting set.  In the old code, this
was covered by the fact that until storage for the serial state was
created for a thread at the time (if any) that we first explicitly set
it, chpl_get_serial() would just return false.  In the new code, thread
private storage is created right away for every thread, so we need to
set the initial serial state explicitly.

The other problem was that we locked the report_lock mutex before we
created it.  Before we did chpl_mutex_init(&report_lock), we called
initChplThreads_generic(), which called chpl_thread_init_generic(),
which called initializeLockReportForThread(), which tried to lock
report_lock and at least on the Mac, this failed.  (It didn't fail on
x86; either that pthreads implementation is less picky or, more likely,
the initial zeroed contents of the lock were good enough.)  Moving the
lock init call before the initChplThreads_generic() call fixes this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15929 3a8e244f-b0f2-452b-bcba-4c88e055c3ca