Improving support for specifying C header files
Prior to this commit, users could specify C header files on the chpl
command line and we would add #includes for them in the generated C
code.  However, the compiler tried to open the header files, requiring
them to be in the current working directory or specified with a path.

Motivated by something Jacob was trying to do last week, this check-in
improves this capability by adding a -I option to the Chapel compiler
(to complement the existing -L support) which will be fed to the
back-end C compiler.  I also disabled the check to see whether the
header file exists so that it could be along the back-end C compiler's
standard header include path.

Added a test to lock in this capability and updated the man page to
document the -I flag.

As a next step, we should unify the way we're dealing with the -I,
-L, and -M flags and their respective environment variables.  In 
particular:

- -I and -M use a vector to store their directories, -L does not
- -L has a normal environment variable version of its setting which
  isn't consistent with -M where the environment variable was distinct
  since it can specify a full colon-separated path rather than just
  a single directory at a time.
- generally speaking the terms path and directory aren't used as 
  consistently in the --help message, man page, and code as one might
  like.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15942 3a8e244f-b0f2-452b-bcba-4c88e055c3ca