Disabled execution of the type expression code in the default wrapper for default constructors if the argument is neither an array nor a domain.  Executing this code results in a leak that cannot be handled. For example, given a class C, if class D is defined as
  class D {
    var x = new C();
  }

then the default constructor used to create a new instance of class C
based on the code in the init expression (copied to the type
expression as if written "var x: (new C()).type = new C();") even if
an actual value was passed to the default constructor.  This needs to
happen for arrays and domains (arguably all value types?).  The
constructor story ought to be reassessed at some point.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15943 3a8e244f-b0f2-452b-bcba-4c88e055c3ca