Fixing broken links in --html view
It seems that at some point our --html view option stopped producing
an HTML file for the "Program" module which made hyperlinks leading
to that module broken.  This may have been due to my recent module
changes, but I'm not certain.  Fixed this and simplified the generation
of the list of modules by using the allModules vector rather than
creating a vector of module symbols on each call to html_view.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15948 3a8e244f-b0f2-452b-bcba-4c88e055c3ca