Removed nested modules from recusive HTML generation
After my change yesterday, when HTML-viewing a module, we would also
HTML-view its nested modules.  In the case of a nested user module,
this would cause it to be generated twice, once on its own and once in
the file of its parent module.  In the case of the program module, it
caused the entire program to be generated twice because the program
module contains all other modules.  This wasted space and made nesting
modules hard to browse.  With this change, we trim the recusive HTML
generation when we encounter nested modules so that nested modules
only appear as top-level modules in the index.html file.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15950 3a8e244f-b0f2-452b-bcba-4c88e055c3ca