Revamped module initialization logic
This commit changes the way we deal with module initialization.
Previously, modules were initialized on their first use by any task.
This required guarding each module initialization with a global
synchronization variable (and for performance reasons, optionally a
per-locale cache of this variable to avoid communication in a case
where we've already determined its been initialized).  This scheme had
some downsides -- primarily the runtime overhead involved and the
complexity of maintaining the module guard code in the compiler.

This new change does all module initialization at program startup time
in the following order:

Consider the directed graph in which an edge from M1 to M2 implies
that module M1 uses module M2.  Starting from the main module, we
initialize the modules in the order implied by a depth-first postorder
search of the graph.  For a given module, all outgoing edges (all
modules that it uses) are initialized in the order they appear in the
program text regardless of dynamic execution order.  A nested module
will ensure that its parents are initialized before it is initialized
but after the modules it uses are initialized.

As a small detail that shouldn't concern the user's view but does
affect our implementation, all modules found through the search of
ChapelStandard are initialized first in the Program module's
initialization routine, then we start tracking memory and allocate the
global endcount (because those things depend on the standard modules
being initialized) and then we initialize all the other modules found
during the search.

This new approach has the benefits of (a) not incurring as much
runtime order, and (b) presenting a reasonably intuitive module
initialization order that is not dependent on things such as order in
which files are listed on the compiler's command line.  For most
well-behaved modules that don't have visible side effects or cyclic
dependences, the order is not expected to matter, but having a
well-defined order gives the programmer something to rely on.

Somewhat surprisingly, only two of our tests needed to be changed as a
result of these new semantics.  In particular, David had two tests
that relied on module initialization occurring when the module was
used.  I've updated these to push the actions that used to be taken in
module initialization into a function and called that function right
after the module uses to reproduce the behavior we were getting
before, but he should review these to make sure he's happy with them
and thinks they continue to have value (I think they do).

The modifications to the individual files are as follows.

I still need to update the documentation and make sure we have the
right existing tests to demonstrate that module initialization occurs
in the right order.


The main code related to generating blocks of module initialization
statements and inserting them into the program's initialization
section and chpl_main:
* compiler/passes/buildDefaultFunctions.cpp

Removed fields related to module guards and added list+set to track
module uses uniquely and in order
* compiler/include/symbol.h
* compiler/AST/symbol.cpp

Removed module guard code and changed our processing of use statements
to add use'd modules to a module symbol's use list
* compiler/passes/scopeResolve.cpp

Squelched the use of the standard module from chapel_code.chpl since
this results in compiling all standard modules into the runtime (which
is inefficient and breaks compilation of chapel_code.c)
* compiler/passes/filesToAST.cpp

Removed code related to module guards:
* runtime/include/chplrt.h
* runtime/src/main.c
* compiler/AST/build.cpp
* compiler/passes/cleanup.cpp

Updated tests to reflect that uses no longer initialize a module
(David should review since the spirit of these tests has changed,
though it still checks that an outer module's initialization runs
first).
* test/modules/diten/moduleNotInitialized.chpl
* test/modules/diten/moduleNotInitialized2.chpl

Fixed a USR_FATAL that was passing a pointer that we just proved
was NULL:
* compiler/passes/normalize.cpp




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15954 3a8e244f-b0f2-452b-bcba-4c88e055c3ca