Added support for specifying realmTypes on the compiler command-line
Prior to this change, the types of realms that would be used for
multirealm executions were hardcoded into runtime/src/arg.c as a
(documented) short-term hack to get us through our last deliverable.
This change introduces support for a compiler flag to specify the
realms on which the program should be run, in the form:

  chpl -snumRealms=2 -srealmTypes='rt1 rt2 rt3'

where rt? is a CHPL_PLATFORM string like linux, linux64, darwin, etc.
This change is implemented simply by code generating a function that
has the same signature as the one that used to be hardcoded into
arg.c, so it shouldn't require any changes in the PVM runtime library
itself.

If the number of realms is 1 (either set explicitly or using
-snumRealms=1), the normal CHPL_TARGET_PLATFORM variable is used to
get the realm type if the -srealmTypes flag is not used.  If the
number of realm types specified does not match the number of realms,
an error is generated.  This error is currently generated at codegen
time (i.e., ridiculously late for something that could be caught
earlier), so a future usability change would be to move the check up
just after command line arguments are parsed.

Currently the realmTypes config param is implemented using a string
though eventually it would be nice to support config params that are
arrays of strings or tuples of strings to avoid the need to quote the
list of realm types.

I added four tests of various erroneous cases.

As a small side change, I turned off the codegen_config() routine when
we're compiling the runtime since the runtime shouldn't have any
active configuration variables.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15956 3a8e244f-b0f2-452b-bcba-4c88e055c3ca