Refactored constructors (via default wrappers) to avoid leaking array and domain fields.  This resulted in the need to start copying and destroying iterator records, making this change rather large.  Plugged a number of other leaks along the way.  Details on a file-by-file basis are listed below.
This commit improves memory management, decreasing the aggregated
leaked memory of running all tests from 764MB to 506MB or from 13.7%
to 9.0% (as a percentage of total memory allocated).

Compiler Changes
----------------

build.cpp:

Renamed _loopexpr as _seqloopexpr and _parloopexpr to distinguish
between for and forall loop expressions.  Removed an explicit autoCopy
in the _parloopexpr that is no longer necessary.

Added more "expr temp" flags to temporaries inserted when building the
IR that are necessary in order for the compiler to insert matching
autoCopy and autoDestroy calls, especially now that iterator records
are autoCopied and autoDestroyed.

Rejiggered the forall loop generated for reductions and scans so that
the _toLeader and _toFollower functions are first called on iterator
recoreds, not iterator classes, thus allowing iterator records to be
correctly destroyed within reductions.  TODO: It may be possible to
eliminate (refactor via inlining) the _toLeader and _toFollower
functions defined on iterator classes.

cleanup.cpp:

Changed specialization of _loopexpr to use the new names.

wrappers.cpp:

Refactored how arrays and domains are treated in the default wrapper
for a default constructor to avoid a major leak when constructing
records or classes with array or domain fields, in which we would leak
these fields.

lowerIterators.cpp:

Refactored how _toLeader and _toFollower functions are rewritten in
the compiler to account for passing the iterator record by reference
instead of returning it (now that iterator records are autoCopied and
autoDestroyed).

Added code to rewrite the autoCopy and autoDestroy functions for
iterator records to account for their fields.

functionResolution.cpp:

Modified requiresImplicitDestroy to return true for functions that
return iterator records (except for iterators).

callDestructors.cpp:

Modified the code that inserts destructor calls for value fields in
records and classes to insert an autoDestroy call for fields if the
field has no destructor (as is the case for iterator records).  TODO:
It may be possible to simplify this code by always using the
autoDestroy call.

Module Changes
--------------

ChapelArray.chpl:

Added an unfortunate definition of assignment from arrays to arrays
(previously handled by the assignment from anything to arrays) that
contains a conditional guard in case the right-hand array (b) is
uninitialized.  This can happen if a record has not yet been
initialized and a loop expression is used to define one of its field.
Then when the iterator record is copied, its record field can be
copied, resulting in an assignment of an array that is not yet
initialized.  If we do a better job of initializing arrays before they
are used in constructors (not just setting the _value field to nil)
then we can remove this conditional and assignment definition.

Added "no implicit copy" pragmas to the _toLeader and _toFollower
functions that wrap their respective primitives to avoid an
unnecessary autoCopy.

ChapelBase.chpl:

Plugged a leak in the initCopy function defined over iterators that
returns arrays since variables in the initCopy function are not
destroyed automatically.  Removed an explicit call to autoCopy the
array (since it is not destroyed) and added a pragma "insert auto
destroy" to the domain (since it is otherwise not destroyed).

Added autoCopy and autoDestroy functions for iterator records.  The
bodies of these functions are modified during the callDestructors
compiler pass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15964 3a8e244f-b0f2-452b-bcba-4c88e055c3ca