This change adds preliminary support for enumerated types for hetero-PVM.
This change adds preliminary support for enumerated types with
CHPL_COMM_HETEROGENEOUS with PVM. Since there isn't (and couldn't be)
a pvm_pkenum, we need to send this via pvm_pkbyte (this submit is
preliminary since this won't work for different endianness/bittedness
architectures). The new pack/unpack algorithm is to iterate through
all the field types in the struct we're communicating; if any of these
are an enumeration, then we have to send the whole structure with
pvm_pkbyte. (We'll likely have to add other types to this later, such as
chpl_task_list_p, for one example.) If we pack the struct this way, we
have to skip the normal algorithm (both on the pack and unpack side).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15968 3a8e244f-b0f2-452b-bcba-4c88e055c3ca