One line change removes 63% of remaining leaks!
Plugged a leak in which the elements of an arithmetic array were
leaked whenever the array was resized.  This accounted for 63% of the
remaining leaks, most of which had to do with adding or removing
indices to associative domains, growing or shrinking associative
arrays, and leaking the elements of these associative arrays.  In the
Madness tests, the elements of these associative arrays are records
with fields of arithmetic domain and array type.

This commit improves memory management, decreasing the aggregated
leaked memory of running all tests from 506MB to 184MB or from 9.0% to
3.3% (as a percentage of total memory allocated).  Strings account for
20.9% of the remaining leaked memory.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15974 3a8e244f-b0f2-452b-bcba-4c88e055c3ca