Removing nondeterminism from generic type names
[ghost written by Steve]

For a generic type like "foo" the compiler will generate a new
name for "foo" as we instantiate it with different types and
parameters.  Historically, the precise name of foo has been
nondeterministic, such that if foo was generic across int32
and real, the name could end up either "foo_int32_real" or
"foo_real_int32".  This has led to a fair amount of cruft in
diffing generated code across versions of the compiler,
different sets of flags, or different architectures.  The
reason for this nondeterminism is that we have stored the
generic substitutions for a type as a hash table based on
a pointer value and iterate over that hash table when generating
the symbol name.  As pointer values change, so does the order
of the entries in the hash table.

This change makes the naming deterministic by using the formal
arguments of the constructor function for the type as an
ordering constraint.  So, rather than iterating over the hash
table, we iterate over the formals, look them up in the hash
table, and generate the name in that order.  These constructor
formals are typically in the order that the member types/params
appear in the class definition, so this also provides a more
useful type name in that it can be read and matched to the
generic members.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15975 3a8e244f-b0f2-452b-bcba-4c88e055c3ca