Add some tests showing odd behavior when promoting arrays with different ranks together.
These tests add 2D and 3D arrays with the same total number of elements
together and sum reduce the result. In the .future test, the 2D array is on
the left and the compiler chooses to use leader/follower iterators, but gets
a compile time tuple index out of bounds error in one of the followers.

In the test without a .future, the 3D array is on the left, and a serial
iterator is used instead. The serial iterator linearizes the arrays and the
sum is computed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15978 3a8e244f-b0f2-452b-bcba-4c88e055c3ca