This change sorts the output of the hetero-PVM generated code types.
This change sorts the output of the hetero-PVM generated code types. It does
this using qsort on the cnames of the types to be added into
_type_structure.c. Before the change, using hetero_PVM would fail across
machines of different architectures since the addresses of where the types
resided (key to hash) would differ. After this change, a vector would
remain similar across all architectures.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15982 3a8e244f-b0f2-452b-bcba-4c88e055c3ca