Fixing GASNet-specific regressions that I've introduced recently
test_module_use_locales is a test whose output changed when the module
initialization order changed, but I didn't catch it because it's
skipped if we don't have multiple locales.  The original point of the
test is now a bit moot (since use clauses don't result in module
initializers being called), but I'll leave the question about whether
or not to remove it to Steve.  I think it still has some value.

noLocalesPerRealm has failed with GASNet testing since I added it
because we added the -nl flag which doesn't make sense for a
multirealm execution.  Setting the numlocales to 0 squelches this flag
and causes it to work as expected.

Fixed a typo on the regressions file.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15984 3a8e244f-b0f2-452b-bcba-4c88e055c3ca