Adding future based on config param setting bug submitted this morning
This future captures the bug from LiWei this morning that shows
that our configuration parameter parsing is not as strict as
our configuration constant/variable parsing.

I've been wondering increasingly lately whether we would want to
rearchitect our sources to make it easier to share sources between the
compiler and runtime.  This is an example where it would be helpful
and lead to more consistency -- not necessarily the argument-parsing
capability (though that might be nice as well), but the various
string-to-type routines which received a fair amount of attention in
the config var/const case.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15989 3a8e244f-b0f2-452b-bcba-4c88e055c3ca