Simplify the modules involving distributions and iterators and make the block distribution robust enough to use for hpcc/fft except for the reduction in the verification step. With this change all currently written leader iterators yield rank-tuples of ranges. Previously, there were leader iterators that yielded ranges, tuples of ranges, and domains, so follower iterators had to have code for each of these cases.
Updated leaders/followers in tests that didn't use this pattern but zip with
iterators that now do, for example the ra-randstream follower.

Changed the iteratorIndex and _getIterator functions in ChapelArray so that
they can handle one-tuple based index sets.

Added the ability to give Block domains strides and use the strides when
iterating over them.

Added a "tuple" function that takes a variable number of arguments and returns
them as a tuple, including one-tuples.

Added a "linearize" iterator to ChapelArray and used it on the 2D part of
some sudoku tests that zip together promoted 1D arrays with promoted 2D arrays.

The hpcc/2008/ra.chpl test is not passing because the RAStream follower
iterator doesn't expect a tuple. Should the test be updated or should a .notest
file be added?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15990 3a8e244f-b0f2-452b-bcba-4c88e055c3ca