Improved module initialization order
Based on discussion at this morning's meeting, changed the BFS for the
module initialization order to start with a main module's parent
module (if one exists) and then crawl through all of its uses rather
than the reverse which is what we did before this checkin.  As
expected, we don't use nested modules and/or modules whose
initialization order matters often enough that this actually affected
the semantics of any existing programs.

I did expand my modInitOrder test to add some other cases that I'd
thought of recently and to update its output (it is the one test that
changed).  For anyone needing a study break (or wanting to
double-check me), I'd recommend reading through the source, predicting
the output, and then checking your (my) answer to see if it's as you
would expect.  Let me know if you think it's wrong in any way.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15996 3a8e244f-b0f2-452b-bcba-4c88e055c3ca