Plugged a leak related to reindexing in which the autoDestroy call inserted for the reindex temporary would not be executed if the function called an explicit return because the autoDestroy call was just inserted at the end of the function (during normalization). Instead of inserting matching autoCopy and autoDestroy calls explicitly, added the "expr temp" flag to this temporary so that it is handled in the usual way.  In so doing this, removed a conditional guard in which we would not reindex if a nil array were passed to the function.  This case no longer seems necessary and is a very nice simplification.
Plugged a leak related to arguments of "in" or "inout" intent in which
they were never destroyed.  Added the "insert auto destroy" flag to
formal temps of "in" intent.

Plugged a leak related to opaque arrays in which the associative
domain and array classes used to implement it were never freed.  In
addition, I made it so that the opaque indices are freed when the
opaque domain is freed.

Plugged a leak related to a global instance of the SeedGeneratorClass
allocated in the initialization of the Random module.  This class had
a single method and no state.  Turned it into a record so that it does
not need to be deallocated.

Inserted code to free the RandomStream class in all tests that use it.

This commit improves memory management, decreasing the aggregated
leaked memory of running all tests from 62MB to 49MB or from 1.1% to
0.87% (as a percentage of total memory allocated).  Strings account
for 78.7% of the remaining leaked memory (38MB).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15999 3a8e244f-b0f2-452b-bcba-4c88e055c3ca