For the most part this is just a reorganization and tidying-up mod.
Reorder file contents so that the function declarations in the .h files
and the definitions in the .c files are in the same order, and are
divided into separate groups of those which have to do with tasks and
those which have to do with threads.

Make the following function name changes, reflecting the fact that the
functions have more to do with initializing and shutting down task
management than threading support:

    initChplThreads -> chpl_tasking_init
    exitChplThreads -> chpl_tasking_exit

Change "thread" to "task" in several places in comments and file-static
symbol names.

Remove chpl_thread_init().  It was misnamed, since it was called first
thing in each compiler-emitted task wrapper function.  We now do the
equivalent thing just before each call to the task wrapper in the task
management code.  This is functionally equivalent, and simplifies both
the interface and the control flow.

Remove the chpl_threads_initialized global variable which was set in
main.c and queried in chplexit.c.  Instead, maintain the knowledge as to
whether or not task management has been set up inside task management
itself, and make the shutdown function chpl_tasking_exit() do nothing if
task management hasn't been set up.

The compiler changes are due to removing chpl_thread_init().



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16006 3a8e244f-b0f2-452b-bcba-4c88e055c3ca