Semantic future about where resolution
Adding a semantic future about where resolution as we discussed
at lunch today.  The argument is that a function with no where
clause should be considered less attractive than a function
with an identical argument list and a where clause because
otherwise the functions with the where clauses *cannot* be
called making them useless and only a source of ambiguities.

Albert recently ran into this when a new array assignment case
was added with no where clause, motivating this future.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16008 3a8e244f-b0f2-452b-bcba-4c88e055c3ca