Merging Block1D and Block distributions to get rid of Block1D
This change combines the Block1D and Block distributions since the
(multidimensional) Block distribution is the one we want to carry
forward and was seen to generate competitive performance at some
point in the past.  Going forward, "BlockDist" is the module to use
for the "Block" distribution.  (i.e., I svn rm-d BlockDist.chpl
and then svn mv-d MultiBlockDist.chpl to BlockDist.chpl).

After this mv, the main change to BlockDist.chpl is that I implemented
the writeThis() methods for the global and local block distributions
more similarly to what we did for the Block1D case because it more
accurately shows which actual locales own each chunk of indices rather
than which locale index owns each (e.g., if one were to oversubscribe
locales, the new scheme would print out 0 1 0 1 whereas the previous
would print out 0 1 2 3.  I did this without re-introducing the locale
ID into the local Block class (which Steve eliminated based on a todo
in his work making a multidimensional distribution).

I also took out the default value of "1" for rank in the BlockDist
as I don't think it makes sense to have a default rank for a
multidimensional distribution (Steve agreed).

Most of the test updates to the testing system involved (a) changing
a use of MultiBlockDist to BlockDist and (b) changing "new Block1D("
into "new Block1D(rank=1, ".  There were also a few changes to tests
relating to debug printing because they printed out Block1D in their
output.


The other main change (which is a very minor one) was to 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16010 3a8e244f-b0f2-452b-bcba-4c88e055c3ca