Multiple main improvements
Based on discussions at the language spec meeting yesterday,
I added three tests that determine what happens if a user
module that is not specified on the command line contains
a main function in the presence/absence of main functions
in the modules defined on the command line.  It turned out
I had some cases working as we wanted, but others not (in
particular, if the main (specified) modules did not define
main but the user (module path) modules did, I was using
the main in the user modules.

Changed buildDefaultFunctions to implement the behavior we
agreed upon and also improved its error message relating
to having an ambiguous error message to print out all
candidate main functions (previously only two were) and
to point to the --main-module flag.

Updated some tests based on these changes -- primarily
their error messages in the case of ambiguous mains.
There was only one test that relied on a main in a
user module which was uts_rec_seq.chpl (a test that
included another module as a means of compiling it
in a different way -- this would be better handled
by Sung's "multiple compopts lines" change anyway).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16018 3a8e244f-b0f2-452b-bcba-4c88e055c3ca