Add a future that has the same basic problem as: multilocale/deitz/oneOrMoreLocales/test_reduce_of_promote_of_distributed.chpl but doesn't use distributed arrays or reductions.
It has an iterator that yields values from inside a coforall loop zipped with
another copy of itself. The label that is inserted after yield points is
defined in the coforall function, but the goto at the beginning of the iterator
is in the advance method. Since there are no uses of the label within the
coforall function, the definition is removed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16021 3a8e244f-b0f2-452b-bcba-4c88e055c3ca