Added an 'exported' function called chpl__autoDestroyRuntimeGlobals if --runtime is used.  This function contains the same contents that chpl__autoDestroyGlobals does for other programs and allows the runtime to call into this function to destroy all data created in chapel_code.chpl.  Added a call to this function after chpl_tasking_exit is called if _runInGDB() returns false (because otherwise we have not initialized any runtime globals since we only called into GDB and are now exiting without executing any of the Chapel code).
Moved the point in compilation at which chpl__autoDestroyGlobals and
chpl__autoDestroyRuntimeGlobals are created to after function
resolution so that we don't (problematically) try to destroy global
variables in modules that are not used.  It is a problem because we
don't know their types.

Moved the point at which the memory report is generated to after the
point at which the threading layer shutdown routine
(chpl_tasking_exit) is called and after the point at which the globals
in chapel_code.chpl are destroyed.  This removes all reported leaks of
"thread private data" which is cleaned up in chpl_tasking_exit.

Added a destructor to the implementing class of single variables in
order to plug a leak in which the threading-layer specific storage
would not be destroyed.  Added a primitive to handle the destruction
of single variables symmetric to sync variables.  Added definitions of
the single variable destruction routines to the MTA and NONE thread
runtimes.  This was already there for pThreads, but was not used.

Added explicit delete calls to 8 tests.

This commit improves memory management.  The aggregated leaked memory
of running all tests is now 42.8MB or 0.76% of total memory allocated.
Strings account for 89.8% of the remaining leaked memory.  The top
three reported leak categories are all string related.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16022 3a8e244f-b0f2-452b-bcba-4c88e055c3ca