Strange leader/follower bug
This program shows a strange problem with zippered iteration over
domains of different index types.  Something in the compiler-generated
code seems to want them to be the same.  I haven't traced this through
to its core cause yet, and may require some help from Steve (the
callchain where the error is triggered seems to be a long chain of
toFollower -> toFollowerHelp -> toFollower -> toFollower -> toFollower
-> toFollower -> ... that I'm not sure how to unravel).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16024 3a8e244f-b0f2-452b-bcba-4c88e055c3ca