Added support for string literal paramaters in classes and records.
The type name includes the full string literal, while the c name includes 
a sanitized, truncated version of the literal. I've included a small 
test as well.

I also switched a couple uses of sprintf and strcat with statically-sized 
buffers to snprintf and strncat, since I was already in the code.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16027 3a8e244f-b0f2-452b-bcba-4c88e055c3ca