Start of a Block-Cyclic distribution
This is a snapshot of the start of a Block-Cyclic distribution.  I've
only implemented the global and local distribution classes so far and
started with the Block distribution, so there's currently a lot of
replicated code that does not relate to Block-Cyclic.  Also added a
pair of tests that lock in what I've got so far.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16029 3a8e244f-b0f2-452b-bcba-4c88e055c3ca