Fixed a leak in the code that handles capturing of the iterator into an array in which the elements of the iterator/array were leaked.
Fixed a leak in forall statements, forall expressions, and reductions
in which the indices yielded by the leader iterator were leaked.
Since leader iterators mostly return tuples of ranges, this was not a
problem, however, in one Madness test, the leader iterator yielded a
record which had fields of array/domain types.

This check fixes all leaks in
  studies/madness/aniruddha/madchap/recordBug/test_likepy
which was the number two largest leak (at 1.49MB not including
strings).

This commit improves memory management.  The aggregated leaked memory
of running all tests is now 41.5MB or 0.74% of total memory allocated.
Strings account for 92.6% of the remaining leaked memory.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16033 3a8e244f-b0f2-452b-bcba-4c88e055c3ca